feat: EXPIRETIME, PEXPIRETIME, SMOVE, and SINTERCARD (#315)
adds four new Redis-compatible commands:

- EXPIRETIME key — returns absolute Unix timestamp (seconds) when the key
  expires; -1 if no expiry, -2 if the key doesn't exist
- PEXPIRETIME key — same but in milliseconds
- SMOVE src dst member — atomically moves a member between sets; handles
  cross-shard keys by doing a srem + sadd across shards
- SINTERCARD numkeys key... [LIMIT n] — returns the cardinality of the
  intersection of N sets with an optional cap; fetches members from each
  key's owning shard before computing the intersection in the execute layer

the internal expiry clock is monotonic (ms since process start), so a new
monotonic_to_unix_ms() helper in time.rs anchors it to wall-clock time once
and uses fast arithmetic on every subsequent call.

AOF persistence: SMOVE on a single shard writes two records (SRem + SAdd)
which replay cleanly; cross-shard SMOVE persists each sub-command to its
own shard's AOF log.

12 integration tests, unit tests for smove and sintercard in keyspace/set.rs.

+++ crates/ember-server/src/metrics.rs
@@ -267,8 +267,8 @@ pub fn spawn_stats_poller(engine: Engine, ctx: Arc<ServerContext>, poll_interval
 
                     // expired and evicted are cumulative totals from shards —
                     // publish the delta so prometheus sees a proper counter.
-                    let expired = total.keys_expired as u64;
-                    let evicted = total.keys_evicted as u64;
+                    let expired = total.keys_expired;
+                    let evicted = total.keys_evicted;
                     let delta_expired = expired.saturating_sub(last_expired);
                     let delta_evicted = evicted.saturating_sub(last_evicted);
                     if delta_expired > 0 {
